refactor: Move i18n into own file
