+++ src/js/config/configuration.ts
@@ -1,6 +1,7 @@
 import equal from "fast-deep-equal";
 import defaultGames from "./sideMenuGames";
-import { addChangeListener, localStorageClear, localStorageGet, localStorageSet, i18n, setI18nLocale, getI18nDefaultLocale, storageClear, storageGet, storageSet } from "../utils/browser";
+import { addChangeListener, localStorageClear, localStorageGet, localStorageSet, storageClear, storageGet, storageSet } from "../utils/browser";
+import { i18n, setI18nLocale, getI18nDefaultLocale } from "../utils/browser/i18n";
 import { ADVANCED_HOME_STYLE, COLORFUL_TABLES, DEF_HOME_HTML } from "./configuration.constants";
 
 export interface Game {
