+++ src/js/ui/LoginButton.jsx
@@ -1,7 +1,7 @@
 // @flow
 import { Button } from "./base/Button";
 import { bgaUrl } from "../utils/constants";
-import { i18n } from "../utils/browser";
+import { i18n } from "../utils/browser/i18n";
 
 export function LoginButton(): React$Element<typeof Button> {
 	return (
