+++ src/js/ui/Table.jsx
@@ -1,7 +1,7 @@
 // @flow
 import type { TableId } from "../types/bga/Table";
 
-import { i18n } from "../utils/browser";
+import { i18n } from "../utils/browser/i18n";
 import { PlayerList } from "./PlayerList";
 import { TableContent } from "./TableContent";
 import { TableFooter } from "./TableFooter";
