+++ src/js/ui/views/OptionsView.tsx
@@ -6,7 +6,8 @@ import Configuration, { AdvancedHomeConfig, HomeConfig, InProgressConfig } from 
 import Switch from "../base/Switch";
 import { Button } from "../base/Button";
 import { isSoundCustom, playMp3, removeCustomMp3, uploadCustomMp3 } from "../../utils/misc/mp3";
-import { getExtensionVersion, i18n, isFirefox } from "../../utils/browser";
+import { getExtensionVersion, isFirefox } from "../../utils/browser";
+import { i18n } from "../../utils/browser/i18n";
 
 type Props = {
   config: Configuration,
